Connect to an SQLite database

Wrapper function to return a SQLiteConnection object for local development.

connect_sqlite(auto_disconnect = TRUE)

Arguments

auto_disconnect
Should the connection be automatically closed when the src is deleted? Set to TRUE if you initialize the connection the call to src_dbi(). Pass NA to auto-disconnect but print a message when this happens.

Value

Returns an S4 object that inherits from DBIConnection. This object is used to communicate with the database engine. Under the hood, dbConnect() returns an object of class SQLiteConnection. See dbConnect() for more details.

Examples

connect_sqlite()

#> <SQLiteConnection>
#>   Path: :memory:
#>   Extensions: TRUE