Note CRUD module - server function

Server function for the noteCrudMod module.

noteCrudMod(input, output, session, group_column, selected_group,
  group_options, category_options, table_id, db_conn,
  style_options = default_styles())

Arguments

input
Standard shiny input

output
Standard shiny output

session
Standard shiny session

group_column
Column in table to group and filter notes by

selected_group
Currently selected group column value

group_options
Group column row value options

category_options
Category column row value options

table_id
Named list with member 'table' and 'schema' referring to a db table containing notes.

db_conn
database connection cursor

style_options
Optional named list of css styles to apply to note panel elements.

Examples

if(interactive()){
  shiny::callModule(
    module = noteCrudMod,
    id = "paragraph",
    style_options = shiny::reactive({
    list(
      "type" = "bullets",
      "header" = list("color" = "#ccc"),
      "panel" = list("scrollY" = TRUE)
      )
    }),
    group_column = "package",
    selected_group = shiny::reactive("shiny"),
    group_options = c("shiny", "shinyWidgets", "dplyr"),
    category_options = c("modules", "noteCrudMod"),
    table_id = list(table = "scroll_demo", schema = "notes"),
    db_conn = connect_sqlite(auto_disconnect = F)
  )
}